backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
pydantic[email]
rich
asyncio
python-multipart